requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.3.0
opencv-python==4.8.0
Pillow==10.0.0
Flask==3.0.0
Flask-CORS==4.0.0
```
gunicorn==21.2.0

